Update & Create Components and tsx Clean
