+++ src/components/ContactUs.tsx
@@ -9,7 +9,7 @@ const ContactUs = ( ) => (
     </div>
 
     <div className="relative h-auto w-full flex justify-center bg-dark-blue">
-      <div className="flex flex-col md:flex-row justify-center card xl:w-[1300px] lg:h-[840px] ">
+      <div className="flex flex-col-reverse md:flex-row justify-center card xl:w-[1300px] lg:h-[840px] ">
 
         <div className="w-full flex flex-col items-start justify-center object-contain ">
           <img src={bobbleMaskBgLast} className="relative lg:absolute md:-w-[360px] xl:w-auto" alt="Booble background"/>
@@ -21,7 +21,7 @@ const ContactUs = ( ) => (
         </div>
 
         <div className="w-full flex flex-col items-start justify-center object-contain py-7 lg:pl-16"> 
-          <h1 className="font-semibold text-6xl text-white text-left pb-3" id="height-line-title">
+          <h1 className="font-semibold text-5xl sm:text-6xl text-light-yellow text-left pb-3" id="height-line-title-mobile sm:height-line-title">
             See what you can do in one week with us.
           </h1>
           <p className="text-medium-blue text-lg text-left pb-5 ">
