+++ src/components/Footer.tsx
@@ -1,0 +1,13 @@
+import logoDoSomeThing from '../assets/images/logo.svg'
+
+const Footer = () => (
+  <>
+    <div className="bg-dark-gray flex justify-center content-center" >
+      <a href="https://www.figma.com/design/dDzuNWi0mSH6oIqU11mMDc/Background-TestModel?node-id=0-1&t=1hiQ5kDFhrNPc9xa-0" target="_blank">
+        <img src={logoDoSomeThing} className="logo" alt="Do Some Thing logo" />
+      </a>
+    </div> 
+  </>
+)
+
+export default Footer
