+++ src/components/MyHero.tsx
@@ -5,10 +5,10 @@ const  MyHero = () => (
   <div className="h-auto flex flex-col items-center justify-end bg-dark-blue">
 
     <div className="pt-36 pb-20 object-contain">
-      <h1 className="text-white text-6xl font-normal pb-3">
+      <h1 className="text-white text-5xl sm:text-6xl font-normal pb-3">
         Create some task
       </h1>
-      <h1 className="text-light-blue text-6xl font-semibold pb-3">
+      <h1 className="text-light-blue text-5xl sm:text-6xl font-semibold pb-3">
         More faster them all
       </h1>
       <p className="text-paragh-gray pb-5 w-full px-12 sm:w-[600px]">
@@ -16,7 +16,7 @@ const  MyHero = () => (
       </p>
 
       <div className="inline max:w-[460px] bg-white rounded-3xl py-3"> 
-        <input type="text" className="p-2 mx-3 rounded-3xl w-80" placeholder="Sendo your email" />
+        <input type="text" className="p-2 mx-3 rounded-3xl sm:w-80" placeholder="Sendo your email" />
         <button className="bg-light-blue text-white p-3 rounded-3xl">
           SEND EMAIL
         </button>
